Pushing previous versions of AI to GitHub
I should've done this in the first place but it's better late than never. I placed the start dates at the top to keep track of my progression so far. Currently, I'm working on V3.1. This version will have columns with injury counts added to the dataset and fix the home and away team mismatch I recently found. Oopsie